requirements.txt
flask==1.1.4
flask-sqlalchemy
psycopg2
flask-migrate==2.5.3
flask-script==2.0.6
marshmallow
flask-bcrypt
pyjwt
pytest
coverage
pytest-cov
python-coveralls
gunicorn
python-dotenv
werkzeug==1.0.1